removed AddLayerModal and merged into treeview
