+++ src/stores/index.ts
@@ -8,9 +8,6 @@ export const dynamicLayers = writable([])
 // progress / indicator bar displayed in the drawer for async fetch data
 export const indicatorProgress = writable(false)
 
-// modal is displayed
-export const modalVisible = writable(false)
-
 // layer map list
 export const layerList = writable(<Layer[]>[])
